added unsigned

+++ 0x00-hello_world/6-size.c
@@ -15,10 +15,10 @@ int main(void)
 	long int d;
 	long long e;
 
-	printf("Size of a as int: %lu byte(s)\n", sizeof(a));
-	printf("Size of b as char: %lu byte(s)\n", sizeof(b));
-	printf("Size of c as float: %lu byte(s)\n", sizeof(c));
-	printf("Size of d as long int: %lu byte(s)\n", sizeof(d));
-	printf("Size of e as long long int: %lu byte(s)\n", sizeof(e));
+	printf("Size of a as int: %lu byte(s)\n", (unsigned long)(sizeof(a));
+	printf("Size of b as char: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of c as float: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of d as long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of e as long long int: %lu byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
